# secp256k1: Optimize NAF conversion.