@matheusd 2021-08-05
Commit msg for [Improve scalar mult readability](https://github.com/decred/dcrd/pull/2695/commits/7216d26a59a328dc9ccd3a2c068de7ea446f4ab5) has "might not otherwise **by** obvious"