@davecgh 2021-08-06
I added an additional edge-condition test case for which results in a 257-bit NAF encoding since I noticed that wasn't explicitly tested.